+++ src/options/keys.ts
@@ -1,0 +1,3 @@
+import { InjectionKey } from "vue";
+
+export const OPTIONS_CONTEXT_KEY = Symbol() as InjectionKey<boolean>;